Refactor: searc card 컴폰너트화 #174
