+++ frontend/react-app/src/components/common/search/channelSearchItem/index.styled.ts
