+++ frontend/react-app/src/components/common/search/searchBar/index.styled.ts
