+++ frontend/react-app/src/components/common/search/searchBar/index.tsx
