+++ frontend/react-app/src/components/common/search/userCard/index.styled.ts
